Django Basics/06 Forms Advanced/requirements.txt
asgiref==3.7.2
crispy-bootstrap4==2023.1
django-crispy-forms==2.1
pillow==10.2.0
sqlparse==0.4.4
typing_extensions==4.9.0
Django==5.0.1
